flask/app/requirements.txt
Flask==1.1.1
flask_cors
python-vlc
youtube_dl
requests